[feat] How do I dynamically update a ListView?, https://flutter.dev/docs/get-started/flutter-for/ios-devs#how-do-i-dynamically-update-a-listview
Instead of creating a “ListView”, create a ListView.builder that takes two key parameters: the initial length of the list, and an ItemBuilder function.

The ItemBuilder function is similar to the cellForItemAt delegate method in an iOS table or collection view, as it takes a position, and returns the cell you want rendered at that position.

Finally, but most importantly, notice that the onTap() function doesn’t recreate the list anymore, but instead .adds to it.